Improve the speed of the async tests
