+++ tests/mock_web_api_server/mock_handler.py
@@ -1,14 +1,9 @@
-import asyncio
 import json
 import logging
-from queue import Queue
-import threading
-import time
 from http import HTTPStatus
-from http.server import HTTPServer, SimpleHTTPRequestHandler
-from typing import Type, Optional, Union
-from unittest import TestCase
-from urllib.parse import urlparse, parse_qs, ParseResult
+from http.server import SimpleHTTPRequestHandler
+from typing import Optional
+from urllib.parse import ParseResult, parse_qs, urlparse
 
 INVALID_AUTH = json.dumps(
     {
@@ -208,142 +203,3 @@ class MockHandler(SimpleHTTPRequestHandler):
             if parsed_path and parsed_path.query:
                 request_body = {k: v[0] for k, v in parse_qs(parsed_path.query).items()}
         return request_body
-
-
-class MockServerThread(threading.Thread):
-    def __init__(
-        self, queue: Union[Queue, asyncio.Queue], test: TestCase, handler: Type[SimpleHTTPRequestHandler] = MockHandler
-    ):
-        threading.Thread.__init__(self)
-        self.handler = handler
-        self.test = test
-        self.queue = queue
-
-    def run(self):
-        self.server = HTTPServer(("localhost", 8888), self.handler)
-        self.server.queue = self.queue
-        self.test.mock_received_requests = self.handler.received_requests
-        self.test.server_url = "http://localhost:8888"
-        self.test.host, self.test.port = self.server.socket.getsockname()
-        self.test.server_started.set()  # threading.Event()
-
-        self.test = None
-        try:
-            self.server.serve_forever(0.05)
-        finally:
-            self.server.server_close()
-
-    def stop(self):
-        self.handler.received_requests = {}
-        with self.server.queue.mutex:
-            del self.server.queue
-        self.server.shutdown()
-        self.join()
-
-    def stop_unsafe(self):
-        self.handler.received_requests = {}
-        del self.server.queue
-        self.server.shutdown()
-        self.join()
-
-
-class ReceivedRequests:
-    def __init__(self, queue: Union[Queue, asyncio.Queue]):
-        self.queue = queue
-        self.received_requests = {}
-
-    def get(self, key: str, default: Optional[int] = None) -> Optional[int]:
-        while not self.queue.empty():
-            path = self.queue.get()
-            self.received_requests[path] = self.received_requests.get(path, 0) + 1
-        return self.received_requests.get(key, default)
-
-    async def get_async(self, key: str, default: Optional[int] = None) -> Optional[int]:
-        while not self.queue.empty():
-            path = await self.queue.get()
-            self.received_requests[path] = self.received_requests.get(path, 0) + 1
-        return self.received_requests.get(key, default)
-
-
-def setup_mock_web_api_server(test: TestCase):
-    test.server_started = threading.Event()
-    test.received_requests = ReceivedRequests(Queue())
-    test.thread = MockServerThread(test.received_requests.queue, test)
-    test.thread.start()
-    test.server_started.wait()
-
-
-def cleanup_mock_web_api_server(test: TestCase):
-    test.thread.stop()
-    test.thread = None
-
-
-def assert_received_request_count(test: TestCase, path: str, min_count: int, timeout: float = 1):
-    start_time = time.time()
-    error = None
-    while time.time() - start_time < timeout:
-        try:
-            assert test.received_requests.get(path, 0) == min_count
-            return
-        except Exception as e:
-            error = e
-            # waiting for some requests to be received
-            time.sleep(0.05)
-
-    if error is not None:
-        raise error
-
-
-def assert_auth_test_count(test: TestCase, expected_count: int):
-    assert_received_request_count(test, "/auth.test", expected_count, 0.5)
-
-
-def setup_mock_web_api_server_async(test: TestCase):
-    test.server_started = threading.Event()
-    test.received_requests = ReceivedRequests(asyncio.Queue())
-    test.thread = MockServerThread(test.received_requests.queue, test)
-    test.thread.start()
-    test.server_started.wait()
-
-
-def cleanup_mock_web_api_server_async(test: TestCase):
-    test.thread.stop_unsafe()
-    test.thread = None
-
-
-async def assert_received_request_count_async(test: TestCase, path: str, min_count: int, timeout: float = 1):
-    start_time = time.time()
-    error = None
-    while time.time() - start_time < timeout:
-        try:
-            assert await test.received_requests.get_async(path, 0) == min_count
-            return
-        except Exception as e:
-            error = e
-            # waiting for mock_received_requests updates
-            await asyncio.sleep(0.05)
-
-    if error is not None:
-        raise error
-
-
-async def assert_auth_test_count_async(test: TestCase, expected_count: int):
-    await assert_received_request_count_async(test, "/auth.test", expected_count, 0.5)
-
-
-# async def assert_auth_test_count_async(test: TestCase, expected_count: int):
-#     await asyncio.sleep(0.1)
-#     retry_count = 0
-#     error = None
-#     while retry_count < 3:
-#         try:
-#             test.mock_received_requests.get("/auth.test", 0) == expected_count
-#             break
-#         except Exception as e:
-#             error = e
-#             retry_count += 1
-#             # waiting for mock_received_requests updates
-#             await asyncio.sleep(0.1)
-
-#     if error is not None:
-#         raise error
