+++ tests/scenario_tests_async/test_message_file_share.py
@@ -9,9 +9,9 @@ from slack_sdk.web.async_client import AsyncWebClient
 from slack_bolt.app.async_app import AsyncApp
 from slack_bolt.request.async_request import AsyncBoltRequest
 from tests.mock_web_api_server import (
-    setup_mock_web_api_server,
-    cleanup_mock_web_api_server,
+    cleanup_mock_web_api_server_async,
     assert_auth_test_count_async,
+    setup_mock_web_api_server_async,
 )
 from tests.utils import remove_os_env_temporarily, restore_os_env, get_event_loop
 
@@ -30,11 +30,11 @@ class TestAsyncMessageFileShare:
     def event_loop(self):
         old_os_env = remove_os_env_temporarily()
         try:
-            setup_mock_web_api_server(self)
+            setup_mock_web_api_server_async(self)
             loop = get_event_loop()
             yield loop
             loop.close()
-            cleanup_mock_web_api_server(self)
+            cleanup_mock_web_api_server_async(self)
         finally:
             restore_os_env(old_os_env)
 
@@ -77,7 +77,7 @@ class TestAsyncMessageFileShare:
         assert response.status == 200
 
         await assert_auth_test_count_async(self, 1)
-        await asyncio.sleep(1)  # wait a bit after auto ack()
+        await asyncio.sleep(0.1)  # wait a bit after auto ack()
         assert result["call_count"] == 2
