+++ tests/scenario_tests_async/test_web_client_customization.py
@@ -1,4 +1,3 @@
-import asyncio
 import json
 import os
 from time import time
@@ -12,9 +11,9 @@ from slack_sdk.signature import SignatureVerifier
 from slack_bolt.request.async_request import AsyncBoltRequest
 from slack_bolt.async_app import AsyncApp
 from tests.mock_web_api_server import (
-    setup_mock_web_api_server,
-    cleanup_mock_web_api_server,
+    cleanup_mock_web_api_server_async,
     assert_auth_test_count_async,
+    setup_mock_web_api_server_async,
 )
 from tests.utils import remove_os_env_temporarily, restore_os_env, get_event_loop
 
@@ -33,11 +32,11 @@ class TestWebClientCustomization:
     def event_loop(self):
         old_os_env = remove_os_env_temporarily()
         try:
-            setup_mock_web_api_server(self)
+            setup_mock_web_api_server_async(self)
             loop = get_event_loop()
             yield loop
             loop.close()
-            cleanup_mock_web_api_server(self)
+            cleanup_mock_web_api_server_async(self)
         finally:
             restore_os_env(old_os_env)
