+++ tests/scenario_tests_async/test_workflow_steps.py
@@ -1,4 +1,3 @@
-import asyncio
 import json
 import logging
 from time import time
@@ -17,9 +16,10 @@ from slack_bolt.workflows.step.utilities.async_configure import AsyncConfigure
 from slack_bolt.workflows.step.utilities.async_fail import AsyncFail
 from slack_bolt.workflows.step.utilities.async_update import AsyncUpdate
 from tests.mock_web_api_server import (
-    setup_mock_web_api_server,
-    cleanup_mock_web_api_server,
+    assert_received_request_count_async,
+    cleanup_mock_web_api_server_async,
     assert_auth_test_count_async,
+    setup_mock_web_api_server_async,
 )
 from tests.utils import remove_os_env_temporarily, restore_os_env, get_event_loop
 
@@ -38,11 +38,11 @@ class TestAsyncWorkflowSteps:
     def event_loop(self):
         old_os_env = remove_os_env_temporarily()
         try:
-            setup_mock_web_api_server(self)
+            setup_mock_web_api_server_async(self)
             loop = get_event_loop()
             yield loop
             loop.close()
-            cleanup_mock_web_api_server(self)
+            cleanup_mock_web_api_server_async(self)
         finally:
             restore_os_env(old_os_env)
 
@@ -104,6 +104,7 @@ class TestAsyncWorkflowSteps:
         response = await app.async_dispatch(request)
         assert response.status == 200
         await assert_auth_test_count_async(self, 1)
+        await assert_received_request_count_async(self, "/views.open", 1)
 
         app = self.build_process_before_response_app("copy_review___")
         response = await app.async_dispatch(request)
@@ -142,6 +143,7 @@ class TestAsyncWorkflowSteps:
         response = await app.async_dispatch(request)
         assert response.status == 200
         await assert_auth_test_count_async(self, 1)
+        await assert_received_request_count_async(self, "/workflows.updateStep", 1)
 
         app = self.build_process_before_response_app("copy_review___")
         response = await app.async_dispatch(request)
@@ -161,8 +163,7 @@ class TestAsyncWorkflowSteps:
         response = await app.async_dispatch(request)
         assert response.status == 200
         await assert_auth_test_count_async(self, 1)
-        await asyncio.sleep(0.5)
-        assert self.mock_received_requests["/workflows.stepCompleted"] == 1
+        await assert_received_request_count_async(self, "/workflows.stepCompleted", 1)
 
         app = self.build_app("copy_review___")
         response = await app.async_dispatch(request)
@@ -182,8 +183,7 @@ class TestAsyncWorkflowSteps:
         response = await app.async_dispatch(request)
         assert response.status == 200
         await assert_auth_test_count_async(self, 1)
-        await asyncio.sleep(0.5)
-        assert self.mock_received_requests["/workflows.stepCompleted"] == 1
+        await assert_received_request_count_async(self, "/workflows.stepCompleted", 1)
 
         app = self.build_process_before_response_app("copy_review___")
         response = await app.async_dispatch(request)
