+++ tests/slack_bolt_async/middleware/request_verification/test_request_verification.py
@@ -1,4 +1,3 @@
-import asyncio
 from time import time
 
 import pytest
